@@ -12,12 +12,10 @@ function App() {
 			<div className="container">
 				<Navbar />
 				<br />
-				<Routes>
-				<Route path="/" exact element={<ExercisesList />} />
-				<Route path="/edit/:id" element={<EditExercise />} />
-				<Route path="/create" element={<CreateExercise />} />
-				<Route path="/user" element={<CreateUser />} />
-				</Routes>
+				<Route path="/" exact component={ExercisesList} />
+				<Route path="/edit/:id" component={EditExercise} />
+				<Route path="/create" component={CreateExercise} />
+				<Route path="/user" component={CreateUser} />
 			</div>
 		</Router>
 	);
